Use QVariant's toLongLong() instead of toInt() per Qt documentation.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.09.02@2787 0a995953-517f-4c62-b348-104708d926b2

+++ Common/spot-on-crypt.h
@@ -141,7 +141,7 @@ class spoton_crypt
   QByteArray publicKeyHash(bool *ok);
   QByteArray symmetricKey(void) const;
   QString cipherType(void) const;
-  int publicKeyCount(void) const;
+  qint64 publicKeyCount(void) const;
   void generatePrivatePublicKeys(const int keySize,
 				 const QString &keyType,
 				 QString &error);
