+++ Common/spot-on-misc.h
@@ -79,7 +79,7 @@ class spoton_misc
 				   const QByteArray &sPublicKey,
 				   const qint64 neighborOid,
 				   const QSqlDatabase &db);
-  static int participantCount(const QString &keyType);
+  static qint64 participantCount(const QString &keyType);
   static void cleanupDatabases(void);
   static void correctSettingsContainer(QHash<QString, QVariant> settings);
   static void enableLog(const bool state);
